Specify complete list of modules.
This precompiles the files so we don't get caught out it trying to compile
the files at runtime.  Long term solution though is to make sure they are
in the library zip file so that we never get caught out due to an invalid
timestamp, or we could remove the source .py files also.