Added setup script to build things on windows using py2exe
Added cleanup script.

Changed the way we handle resources files to make it work better when module
files are bundled inside a zipfile.

Changed the converters directory a bit.  Before it was strange because we
would mostly use execfile() to access them (basically treating them like
pluggable modules).  But sometimes we would just use a normal import
statement.  The new system is:
 - Made basic.py is now mvc.simpleconverters.py.  We always import it if we
   want to use it.
 - The other converters are now standalone modules located in
   the resources/converters directory.