bz19460: Fix converter's finalize() function.
Properly clean up after ourselves due to a missing import and some
bad code.

Also made it a bit nicer.  Now we try to clean up the stale file as well,
and if that didn't work, just log and re-raise the original exception
and let caller handle the original error.