Winsparkle functionality
Added code to ship the winsparkle .DLL.

Changed the Application class to inherit from SignalEmitter.  The base class
doesn't do anything with this, but the widgets Application class uses it to
emit a signal when the main window is first shown.  This is used by winsparkle
to know when it can initialize itself (it needs to wait for the main window to
give the winsparkle popup context).

Moved signals module to mvc rather than mvc.widgets.